fix(engine): terminal-warning discipline and atomic override rollback
Only a terminal Warning state continues a flow: the Incomplete +
non-blocking-step fallbacks in run_steps, rerun selection, and the
progress renderer are gone, so a missing tool or process failure can no
longer produce an overall successful run. Workspace run failures write
the terminal manifest status while still holding the workspace lock.
Overwrite rollback is now checked: removal/restore problems are
surfaced in the error reason with the stranded backup path. Config
override rollback restores via atomic replaces instead of truncating
write_bytes. The coverage policy module moved out of the runtime
package into the single test that consumes it, and reconciliation
tests derive the canonical chain from the real builder.
